requirements.txt
anyio==3.6.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
cycler==0.11.0
Cython==0.29.33
fastapi==0.91.0
fonttools==4.38.0
h11==0.14.0
httptools==0.5.0
idna==3.4
imageai==3.0.3
importlib-metadata==6.0.0
iniconfig==2.0.0
kiwisolver==1.4.4
matplotlib==3.5.3
mock==4.0.3
numpy==1.21.6
opencv-python==4.7.0.68
packaging==23.0
Pillow==9.4.0
pluggy==1.0.0
py==1.11.0
pydantic==1.10.4
pyparsing==3.0.9
pytest==7.1.3
python-dateutil==2.8.2
python-dotenv==0.21.1
python-multipart==0.0.5
PyYAML==6.0
requests==2.28.2
scipy==1.7.3
six==1.16.0
sniffio==1.3.0
starlette==0.24.0
tomli==2.0.1
torch==1.13.1 --extra-index-url https://download.pytorch.org/whl/cpu
torchaudio==0.13.1 --extra-index-url https://download.pytorch.org/whl/cpu
torchvision==0.14.1 --extra-index-url https://download.pytorch.org/whl/cpu
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.14
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
zipp==3.13.0
